Modify: attraction.js split into module
