+++ taipei-day-trip/static/index.html
@@ -49,7 +49,7 @@
   <div class="popupbackground"></div>
   <div class="popupbar">
     <div class="popupbar--popup">
-      <form class="popupbar--popup--form" id="signin" onsubmit="signInValidation(event)">
+      <form class="popupbar--popup--form" id="signin">
         <div class="popupbar--popup--form--title">登入會員帳號</div>
         <input type="email" name="inputEmail" id="popupbar--popup--form--email" class="popupbar--popup--form--input" placeholder="輸入電子信箱">
         <input type="password" name="inputPassword" id="popupbar--popup--form--password" class="popupbar--popup--form--input" placeholder="輸入密碼">
